Get Seasons Use case added

+++ app/src/main/java/com/david/f1stats/data/model/rankingDriver/RankingDriverData.kt
@@ -3,7 +3,7 @@ package com.david.f1stats.data.model.rankingDriver
 data class RankingDriverData(
     val behind: Int,
     val driver: RankingDriverInfoData,
-    val points: Int,
+    val points: Float,
     val position: Int,
     val season: Int,
     val team: RankingDriverTeamData,
