+++ app/src/main/java/com/david/f1stats/data/model/rankingTeam/RankingTeamData.kt
@@ -1,7 +1,7 @@
 package com.david.f1stats.data.model.rankingTeam
 
 data class RankingTeamData(
-    val points: Int,
+    val points: Float,
     val position: Int,
     val season: Int,
     val team: RankingTeamInfoData
